renamed fils to hend with '_view' for clearity
