+++ backend/logic/view/progress_bar_view.py
