+++ backend/logic/view/reset_specific_board_view.py
